Stop declaring RecordWriter methods as inline
A future commit will make `RecordWriter` an abstract base class with
multiple implementations to choose from. This means that every call will
necessarily indirect through a vtable, and there will no longer be any
advantage to having declared these methods as inline. In preparation for
that change, move the methods into the .cpp file.

There are no changes to the contents of any of these methods. The only
changes are moving them and removing the `inline` declarations.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>