Make RecordWriter an abstract base class
Keep the utility methods for writing objects of different types to the
sink in the base class, so that they can be shared by different writers.
Make `d_sink` a protected member of the parent class, so that these
utility methods in the parent class have a sink to work with.

Rather than construct instances of subclasses directly, expose
a `createRecordWriter` factory function that creates and configures an
appropriate `RecordWriter` instance based on the provided `FileFormat`.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>