Introduce a one-pass high water mark aggregator
By tracking the last known high water mark as of the last allocation or
deallocation at any given location, we can report records contributing
to the high water mark or records indicating leaked allocations by just
looping over the history associated with each allocation.

Tracking all of the information that we need takes time proportional to
the number of allocations, and reporting the high water mark or leaked
records takes time proportional to the number of distinct allocation
locations.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>