Introduce AggregatingRecordWriter
This implementation of the `RecordWriter` interface processes the
allocations fed to it in-memory in order to detect the high water mark
records or the leaked records. Once all allocations have been processed,
it writes information about either the high water mark records or the
leaked records to the capture file. It does not retain history of each
allocation that was performed, only the number of allocations and number
of bytes that each allocation location contributed to the high water
mark or to the leaked allocations.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>