Ignore allocations under calloc and posix_memalign
Apparently Alpine implements `calloc` in terms of `malloc`, and
`posix_memalign` in terms of `aligned_alloc`. Ensure that we don't
record duplicate allocations on Alpine.

Our `ALL_ALLOCATIONS` capture file format incorrectly recorded both the
`calloc` and the `malloc` it is implemented in terms of, but because the
`malloc` happens first and returns the same pointer as the `calloc`
later returns, the aggregator wound up overwriting the allocation record
for the `malloc` with the one for the `calloc`, silently dropping it.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>