tests: Exercise the `HighWaterMarkAggregator`
Add a test helper to `memray._memray` to help us exercise
`HighWaterMarkAggregator`. I originally intended to put this in
`_memray_test_utils.pyx`, but because that doesn't link in any `.cpp`
files or `libunwind`, that turned out to require a lot of changes, and
it was easier to just put this in `_memray.pyx`.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>