ci: Automatically publish docs for new releases
On each new release, automatically build our docs and publish them to
GitHub Pages. The docs are also built by our linting workflow, but we
don't attempt to reuse that build here, because the linting build
includes a phony 99.99 version number that it uses for exercising the
news fragments in the news/ directory.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>