tests: Handle incomplete files from process pools
When Memray is used in `--follow-fork` mode and traces into a child
process owned by a process pool, the capture file for that child process
will never be cleanly closed, because the pool forcefully kills its
worker processes once it runs out of work for them to do.

When a Memray capture file is missing its trailer record, we truncate
the last record in the file, rather than process it with potentially
invalid contents (because we can't distinguish between zero bytes that
are part of a valid record versus zero bytes that are padding created
when we grow the capture file).

Because of this, we need to be prepared to handle up to one missing
record per worker process in the process pool. The last record written
by each worker for this test will always be a free, so we need to
account for up to one missing free per worker.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>