tests: Make `PYTHONMALLOC=malloc` optional
Since `PYTHONMALLOC=malloc` is not usable in free-threaded builds, we
need to update the test suite to stop relying on it. For tests in the
test suite specifically testing what happens when `PYTHONMALLOC=malloc`
is used, skip those tests if the interpreter doesn't support it. For
tests where `PYTHONMALLOC=malloc` was being used while testing something
else, stop setting `PYTHONMALLOC=malloc` and update the test to pass
without it.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>