Fix top allocation locations in stats report
This has been subtly broken for a while, but the changes to compute line
numbers on the reader side instead of the writer side broke it
considerably worse. The stats reporter wants to get a unique id
representing the file/function/line where the allocation occurred, so
that it can track which locations are the heaviest allocators. We've
been using a `frame_id_t` to track that. For Python 3.10 and lower,
that works fine, but for Python 3.11 in `--native` mode, the same
file/function/line can have two different frame IDs for the two
different possible `is_entry` flags. And worse, after the changes to
compute line numbers on the reader side, the frame ID is now unique per
file/function/instruction rather than file/function/line, meaning that
two different function calls on the same line of Python source code will
be tracked by the `stats` reporter as two different allocation
locations.

Correct this by introducing a new concept, a `location_id_t`, which
uniquely identifies a file/function/line tuple.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>